requirements.txt
addict==2.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
blis==1.3.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
espeakng-loader==0.2.4
filelock==3.20.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
groq==0.33.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kokoro==0.9.4
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
loguru==0.7.3
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
misaki==0.9.4
mpmath==1.3.0
murmurhash==1.0.13
narwhals==2.10.0
networkx==3.5
num2words==0.5.14
numpy==2.3.4
packaging==25.0
pandas==2.3.3
phonemizer-fork==3.3.2
pillow==12.0.0
preshed==3.0.10
protobuf==6.33.0
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
rdflib==7.4.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rfc3986==1.5.0
rich==14.2.0
rpds-py==0.28.0
safetensors==0.6.2
segments==2.3.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
smart_open==7.4.2
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.3
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
thinc==8.3.6
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
webrtcvad==2.0.10
wrapt==2.0.0
